refactor(runtime/llm): extract executeSingleToolCall to own module
Moves the ~530 LOC single-tool dispatch function — the per-tool-call
path that runs every gate between "model emitted tool_use" and "tool
result on history" — out of chat-executor-tool-loop.ts into a new
chat-executor-single-tool-dispatch.ts. The main loop file drops
2,011 -> 1,410 LOC (49% total reduction from 2,761 at session start).

Moved to the new module:
- executeSingleToolCall (the exported function; called twice from the
  main loop for parallel + serial batches)
- extractDiscoveredToolNamesFromSearchResult (file-private helper,
  only used by executeSingleToolCall's post-exec tool-discovery path)

The main loop imports executeSingleToolCall from the new module and
uses it in the same two call sites. All the tool-utility / hook /
envelope / protocol / budget / routing / workspace-policy imports
that executeSingleToolCall needed are now local to the new module.

Main loop imports cleaned up (no longer needed after extraction):
- LLMToolCall, ToolCallAction (types)
- isRuntimeLimitExceeded + checkToolCallPermission +
  normalizeToolCallArguments + repairToolCallArgumentsFromMessageText
  + parseToolCallArguments + executeToolWithRetry +
  summarizeToolArgumentChanges (from chat-executor-tool-utils)
- pushToolResultMessage + syncToolProtocolSnapshot (from
  chat-executor-tool-protocol-helpers)
- canonicalizeExplicitArtifactReferenceArgs +
  enforceTopLevelExecutionEnvelope (from envelope helpers)
- preflightStaleCopiedCmakeHarnessInvocation (from recovery helpers)
- ANTI_FABRICATION_HARNESS_OVERWRITE_REASON +
  evaluateWriteOverFailedVerification (verification-target-guard)
- evaluateShellWorkspaceWritePolicy (shell-write-policy)
- buildPromptToolContent (chat-executor-text)
- dispatchHooks + defaultHookExecutor (hooks/index)
- applyToolResultBudget (tool-result-budget)
- recordToolProtocolResult (tool-protocol-state)

Purely file-organization — zero behavior delta. Typecheck green.
llm/ test suite (720 tests, 1 pre-existing token-budget compaction
failure unchanged from main) matches baseline.
